requirement.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
MySQL-python==1.2.5
beautifulsoup4==4.3.2
lxml==3.4.3